Combustible refactoring
