+++ groovy/postInit/chains/Chemistry/LeadChain.groovy
@@ -198,17 +198,7 @@ class Combustible {
     }
 }
 
-def combustibles = [
-        new Combustible('dustCarbon', 1, 1, 'dustTinyDarkAsh'),
-        new Combustible('gemCoke', 1, 3, 'dustTinyAsh'),
-        new Combustible('dustCoke', 1, 3, 'dustTinyAsh'),
-        new Combustible('gemAnthracite', 1, 2, 'dustTinyAsh'),
-        new Combustible('dustAnthracite', 1, 2, 'dustTinyAsh'),
-        new Combustible('gemCoal', 2, 4),
-        new Combustible('dustCoal', 2, 4),
-        new Combustible('gemCharcoal', 2, 4),
-        new Combustible('dustCharcoal', 2, 4)
-]
+def combustibles = Globals.combustibles
 
 for (combustible in combustibles) {
     EBF.recipeBuilder()
