+++ groovy/postInit/chains/Chemistry/TinChain.groovy
@@ -69,35 +69,12 @@ CHEMICAL_BATH.recipeBuilder()
         .buildAndRegister()
 
 //SMELTING
-class Combustible {
-    String name
-    String byproduct
-    int amount_required
-    int duration
-    Combustible(name, amount_required, duration, byproduct = 'dustTinyDarkAsh') {
-        this.name = name
-        this.amount_required = amount_required
-        this.duration = duration
-        this.byproduct = byproduct
-    }
-}
-
-def combustibles = [
-        new Combustible('gemCoke', 1, 3, 'dustTinyAsh'),
-        new Combustible('dustCoke', 1, 3, 'dustTinyAsh'),
-        new Combustible('gemAnthracite', 1, 2, 'dustTinyAsh'),
-        new Combustible('dustAnthracite', 1, 2, 'dustTinyAsh'),
-        new Combustible('gemCoal', 2, 4),
-        new Combustible('dustCoal', 2, 4),
-        new Combustible('gemCharcoal', 2, 4),
-        new Combustible('dustCharcoal', 2, 4),
-        new Combustible('dustCarbon', 1, 1, 'dustTinyDarkAsh')
-]
+def combustibles = Globals.combustibles
 
 for (combustible in combustibles) {
     REACTION_FURNACE.recipeBuilder()
             .inputs(ore('dustCassiteriteConcentrate'))
-            .inputs(ore(combustible.name) * (combustible.amount_required))
+            .inputs(ore(combustible.name) * (combustible.amount_required) * 2)
             .outputs(metaitem(combustible.byproduct))
             .fluidOutputs(fluid('crude_tin') * 190)
             .fluidOutputs(fluid('tin_dross') * 25)
@@ -108,7 +85,7 @@ for (combustible in combustibles) {
 
     REACTION_FURNACE.recipeBuilder()
             .inputs(ore('dustLeachedCassiteriteConcentrate'))
-            .inputs(ore(combustible.name) * (combustible.amount_required))
+            .inputs(ore(combustible.name) * (combustible.amount_required) * 2)
             .outputs(metaitem(combustible.byproduct))
             .fluidOutputs(fluid('crude_tin') * 190)
             .fluidOutputs(fluid('tin_dross') * 25)
