+++ groovy/postInit/chains/IronChain.groovy
@@ -1,6 +1,5 @@
 import static globals.Globals.*
 
-
 class Blastable {
     String name
     int amount_required
@@ -31,19 +30,6 @@ class Reductant {
     }
 }
 
-class Combustible {
-    String name
-    String byproduct
-    int amount_required
-    int duration
-    Combustible(name, amount_required, duration, byproduct = 'dustTinyDarkAsh') {
-        this.name = name
-        this.amount_required = amount_required
-        this.duration = duration
-        this.byproduct = byproduct
-    }
-}
-
 def PBF_RECIPES = recipemap("primitive_blast_furnace")
 def EBF_RECIPES = recipemap("electric_blast_furnace")
 
@@ -67,18 +53,7 @@ def reductants = [
     new Reductant('hydrogen', 'steam', 2, 1)
 ]
 
-def combustibles = [
-    new Combustible('gemCoke', 1, 3, 'dustTinyAsh'),
-    new Combustible('dustCoke', 1, 3, 'dustTinyAsh'),
-    new Combustible('gemAnthracite', 1, 2, 'dustTinyAsh'),
-    new Combustible('dustAnthracite', 1, 2, 'dustTinyAsh'),
-    new Combustible('gemCoal', 2, 4),
-    new Combustible('dustCoal', 2, 4),
-    new Combustible('gemCharcoal', 2, 4),
-    new Combustible('dustCharcoal', 2, 4),
-    new Combustible('dustCarbon', 1, 1, 'dustTinyDarkAsh')
-]
-
+def combustibles = Globals.combustibles
 
 for (blastable in blastables) {
     for (combustible in combustibles) {
